Port `transform` tests to new integration test style
